refact: Player 컴포넌트 분리
